[RENDERER] Can see characters as squares

+++ Stealth/game/src/entity/Guard.cpp
@@ -13,7 +13,7 @@
 
 using namespace std;
 
-Guard::Guard(const string &name, Point3d const &pos) : Character(name, pos)
+Guard::Guard(const string &name, Point3f const &pos) : Character(name, pos)
 {
     controller = std::unique_ptr<Controller>(new AIController());
 }
